requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Image processing
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3

# OCR
pytesseract==0.3.10
onnxruntime>=1.16.0
onnxtr>=0.2.0
python-doctr[torch]>=0.6.0

# PDF processing
pdf2image==1.16.3
PyMuPDF>=1.23.0

# Machine Learning
scikit-learn==1.3.2
torch==2.1.0
torchvision==0.16.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
tqdm==4.66.1
PyYAML==6.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code quality
black==23.11.0
flake8==6.1.0

